extracting out uploader content
