+++ components/index.js
@@ -16,3 +16,4 @@ export { default as BoldTitleAndValue } from './BoldTitleAndValue'
 export { default as NextLink } from './NextLink'
 export { default as FutureStack } from './FutureStack'
 export { default as ProjectWrapper } from './ProjectWrapper'
+export { default as JsonDisplay } from './JsonDisplay'
