+++ lib/index.js
@@ -8,3 +8,4 @@ export { default as commify } from './commify'
 export { default as bigNumberToEth } from './bigNumberToEth'
 export { default as bigNumberFrom } from './bigNumberFrom'
 export { default as getEthereumGasEstimate } from './getEthereumGasEstimate'
+export { default as uploaderContent } from './uploaderContent'
